move coloredcobra initialization to command wrapper
